refactor: Plan 클래스 패키지 이동 및 구조 개선
1. Plan 클래스를 entity에서 model 패키지로 이동한다.
2. updateAt 필드를 추가하여 변경 시간을 관리한다.
3. Lombok 어노테이션(@Builder, @AllArgsConstructor)으로 코드 간결성을 높인다.
4. 관련된 import 경로를 수정하여 패키지 이동을 반영한다.

Resolves #22
