+++ src/test/java/com/posicube/assignment/domain/plan/PlanTest.java
@@ -1,8 +1,8 @@
 package com.posicube.assignment.domain.plan;
 
 import com.posicube.assignment.common.exception.BaseException;
-import com.posicube.assignment.plan.domain.entity.Plan;
-import com.posicube.assignment.plan.domain.vo.PlanType;
+import com.posicube.assignment.plan.domain.model.Plan;
+import com.posicube.assignment.plan.domain.model.PlanType;
 import com.posicube.assignment.plan.exception.PlanExceptionStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
